liboping: remove sudo requirement
liboping is opening a raw socket for ping, which requires root privileges, but opening it as a datagram seems to not restrict anything it tries to do and doesn't require root privileges. However, in case there is something it does that needs a raw socket, if it runs as root, it will still open a raw socket.